+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+* ...
+
+
 # Version 0.2.1 [2022-07-25]
 
 ## New Features